Refactor task handling: split handler logic into more separate files. Created new index.js files to consolidate the different handlers (buttons,modals,selects).
