--- events/interaction/features/tasks/handlers/index.js
@@ -1,3 +1,0 @@
-export * from "./buttons.js";
-export * from "./selects.js";
-export * from "./modals.js";
