+++ events/interaction/handlers/button.js
@@ -3,7 +3,7 @@ import { handleClockInButton } from "../features/sessions/handle-clock-in-button
 import { handleClockingInButtons } from "../features/sessions/handle-clock-in-interactions.js";
 import { handleClockOutButton } from "../features/sessions/handle-clock-out-button.js";
 import { handleTimerButtons } from "../features/sessions/handle-timer-buttons.js";
-import { handleTaskButtons } from "../features/tasks/handlers/index.js";
+import { handleTaskButtons } from "../features/tasks/handlers/buttons/index.js";
 import { projectCreateModal } from "../features/projectHelper/project-create-modal.js";
 import { projectEditModal } from "../features/projectHelper/project-edit-modal.js";
 import { projectDeleteConfirmation } from "../features/projectHelper/project-delete.js";
